Move MCP "runtime" logic to `mcp` crate.
