--- app/src/ai/mcp/templatable_manager/utils.rs
@@ -1,83 +1,0 @@
-//! Capability-gating helpers used during MCP server startup.
-//!
-//! Each `query_*_for` function pairs a capability check with the actual list
-//! call from rmcp, gating the call on advertisement and failing soft on errors.
-//! They take the list call as a closure so unit tests can drive the gate-and-
-//! fail-soft control flow with a fake `RunningService` substitute.
-
-/// Whether to query `resources/list` for a server with the given capabilities.
-///
-/// Per the MCP spec, the client should only invoke a list method when the server
-/// has advertised the corresponding capability during initialization.
-pub(super) fn should_query_resources(
-    capabilities: Option<&rmcp::model::ServerCapabilities>,
-) -> bool {
-    capabilities.is_some_and(|c| c.resources.is_some())
-}
-
-/// Whether to query `tools/list` for a server with the given capabilities.
-///
-/// Per the MCP spec, the client should only invoke a list method when the server
-/// has advertised the corresponding capability during initialization.
-pub(super) fn should_query_tools(capabilities: Option<&rmcp::model::ServerCapabilities>) -> bool {
-    capabilities.is_some_and(|c| c.tools.is_some())
-}
-
-/// Query `resources/list` for a connected MCP server.
-///
-/// Skips the call entirely when `resources` was not advertised. Treats any
-/// listing error as "no resources" (fail-soft) so a flaky `resources/list`
-/// does not abort the entire server startup. Mirrors the behavior of
-/// [`query_tools_for`] so the two capabilities are handled symmetrically.
-pub(super) async fn query_resources_for<F, Fut>(
-    capabilities: Option<&rmcp::model::ServerCapabilities>,
-    server_name: &str,
-    list_resources: F,
-) -> Vec<rmcp::model::Resource>
-where
-    F: FnOnce() -> Fut,
-    Fut: std::future::Future<Output = Result<Vec<rmcp::model::Resource>, rmcp::ServiceError>>,
-{
-    if !should_query_resources(capabilities) {
-        return Vec::new();
-    }
-    match list_resources().await {
-        Ok(result) => result,
-        Err(err) => {
-            log::warn!("Failed to list resources for MCP server '{server_name}': {err}");
-            Vec::new()
-        }
-    }
-}
-
-/// Query `tools/list` for a connected MCP server.
-///
-/// Skips the call entirely when `tools` was not advertised. Treats any listing
-/// error as "no tools" (fail-soft) so a transient `tools/list` failure does
-/// not abort the entire server startup — the user-visible regression #6798
-/// was rooted in the prior asymmetric handling, where a tools-list error on
-/// a server with healthy resources would propagate and fail startup.
-pub(super) async fn query_tools_for<F, Fut>(
-    capabilities: Option<&rmcp::model::ServerCapabilities>,
-    server_name: &str,
-    list_tools: F,
-) -> Vec<rmcp::model::Tool>
-where
-    F: FnOnce() -> Fut,
-    Fut: std::future::Future<Output = Result<Vec<rmcp::model::Tool>, rmcp::ServiceError>>,
-{
-    if !should_query_tools(capabilities) {
-        return Vec::new();
-    }
-    match list_tools().await {
-        Ok(result) => result,
-        Err(err) => {
-            log::warn!("Failed to list tools for MCP server '{server_name}': {err}");
-            Vec::new()
-        }
-    }
-}
-
-#[cfg(all(test, not(target_family = "wasm")))]
-#[path = "utils_tests.rs"]
-mod tests;
