+++ crates/mcp/src/lib.rs
@@ -1,1 +1,3 @@
+pub mod oauth;
+pub mod runtime;
 pub mod sse_transport;
