+++ crates/mcp/src/oauth_tests.rs
